refactor(reactjs-todo-davinci): split object value component
