+++ javascript/reactjs-todo-davinci/package.json
@@ -39,7 +39,7 @@
     "webpack-dev-server": "^5.1.0"
   },
   "dependencies": {
-    "@forgerock/davinci-client": "https://pkg.pr.new/@forgerock/davinci-client@638",
+    "@forgerock/davinci-client": "^0.0.0-beta-20260615202501",
     "@forgerock/oidc-client": "latest",
     "@forgerock/sdk-utilities": "latest",
     "@forgerock/protect": "latest",
